backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
pydantic>=1.10.7
neo4j>=5.8.0
requests>=2.31.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
transformers>=4.28.1
torch>=2.0.0
numpy>=1.24.3
pandas>=2.0.1
pdfplumber>=0.9.0
python-docx>=0.8.11
python-pptx>=0.6.21
scikit-learn>=1.2.2
nltk>=3.8.1
jieba>=0.42.1
